requirements-dev.txt
Django>=4.0,<5.0
djangorestframework
django-crispy-forms
psycopg2-binary
openai==0.28
python-docx
PyPDF2
pandas
python-dotenv
textract
requests
beautifulsoup4
python-pptx
openpyxl
python-docx